Lab10/requirements.txt
Django==5.2.12
asgiref==3.11.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
djangorestframework==3.16.0
